requirements.txt
flask==3.0.3
flask-login==0.6.3
psycopg2-binary==2.9.9
bcrypt==4.1.3
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==21.2.0
pdfplumber==0.11.4
